Import LangID fix and filter out "und" result
We did not handle "und" and thought that und is a foreign language
to users as it does not match any of the system languages. Though "und" is
a valid BCP47 tag, but most clients are not aware of this and thus not
handling this. So, filtering out "und" from the result. An empty list
result should effectively mean "und".

Test: No translation action is shown when selecting
     DE19 1234 1234 1234 1234 12  (with a new model file)

Test: Translation action is shown when selecting some foreign text.
      (with a new model file)

BUG: 140542425

Merged-In: I0e1aee3ea6b32b1d06d8ef638091fc2ac3d57763

Change-Id: Ic89e53b64553c02c5dc6fc710e14289c8a692674