Send signal 37 for coverage dump
The workaround was orignated from aosp/1361320.

I can see the coverage dump code is triggered successfully but
I am having a hard time in getting the coverage report locally.
The CL should not have negative impact, so let's get this submitted and
see if the number goes up in the dashboard.


Test: atest libtextclassifier_java_test
Test: atest TextClassifierServiceTest


BUG: 169293677

Change-Id: I74f70105c74fabda590eb37812a2c05c4dfadefb