Add back the workaround for coverage
It was added in ag/12595075 to workaround an issue that some methods
are not considered as tested although they are.

Somehow I accidentally removed it in ag/13332198...

Adding it back and hopefully this gives us back the missing 2%
coverage.

Bug: 179686141
Merged-In: I5bc23741edf1da74fc15a4b25f4b43efe352e88b

Change-Id: Ice399b12ea004acb3c9e441a8e938b080828c60c