Add defensive try-catch to ModelDownloader inside TCS.
We observed some crashes caused by RTE thrown by WorkManager, which is
the major reason for this change. Additionally, we do not want
ModelDownloader crash the ExtServices even if it goes wrong. We will
be able to observe ModelDownloader failures from our logs even if we
catch all those exceptions.

BUG: 219998215
Change-Id: Ic312f46a12dcb6a11b224876517591a89e11d31f