Add force-no-test-error=false to libtextclassifier_tests
This option(newly added) tells GTest to not raise an exception if the test binary is missing.

Currently a missing binary is ignored silently by GTest. But we want to start throwing exceptions when it happens. We anticipated that there will be modules that are misconfigured already and so this option will be used to exempt such modules.

It seems that the module in this change are not configured properly and end up raising exception(WAI) when we enable the feature. so we are exempting them manually from the error. If you have a POC for these module, feel free to share with me and I will create a bug for them to fix it.

This is the context- b/352594845 and https://b.corp.google.com/issues/348446766#comment16

TLDR: The test modules are scheduled to run for both arm64 and arm abi, but they are not compiled for both abi. So the test module fails for the abi that was not compiled.

Bug: 352594845
Test: presubmit
(cherry picked from https://android-review.googlesource.com/q/commit:67fdf947135f34198d4679116b01fbbe21fbbe3a)
Merged-In: I146b3c967fd9ee941cdd4c30934c88fb7e501abc
Change-Id: I146b3c967fd9ee941cdd4c30934c88fb7e501abc